# core/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/core/api/v2/users/{userId}/bills": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть все счета пользователя",
        "operationId": "getUserBillsV2",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientBillDtoV2"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Создать счет для пользователя",
        "description": "Возвращает информацию о созданном счете.",
        "operationId": "createBillV2",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillCreationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBillDtoV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/core/api/v1/users/{userId}/bills": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть все счета пользователя",
        "description": "Описание прав:\n* BILL_READ - доступ к своим счетам\n* BILL_READ_OTHERS - доступ к счетам любого клиента\n",
        "operationId": "getUserBillsV1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      },
      "post": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Создать счет для пользователя",
        "description": "Возвращает информацию о созданном счете.\n\nОписание прав:\n* BILL_WRITE - доступ к открытию своего счета\n* BILL_WRITE_OTHERS - доступ к открытию счета у любого клиента\n",
        "operationId": "createBillV1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      }
    },
    "/users/{userId}/bills": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть все счета пользователя",
        "description": "Описание прав:\n* BILL_READ - доступ к своим счетам\n* BILL_READ_OTHERS - доступ к счетам любого клиента\n",
        "operationId": "getUserBillsV1_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      },
      "post": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Создать счет для пользователя",
        "description": "Возвращает информацию о созданном счете.\n\nОписание прав:\n* BILL_WRITE - доступ к открытию своего счета\n* BILL_WRITE_OTHERS - доступ к открытию счета у любого клиента\n",
        "operationId": "createBillV1_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      }
    },
...
    "/core/api/v2/users/{userId}/bills/{billId}": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть информацию о счете пользователя",
        "operationId": "getUserBillV2",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClientBillDtoV2"
                }
              }
            }
          }
        },
        "deprecated": true
      },
      "delete": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Закрыть счет у пользователя",
        "operationId": "closeUserBillV2",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        },
        "deprecated": true
      }
    },
...
    "/core/api/v2/bills": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть все счета",
        "operationId": "getBillsV2",
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BillDtoV2"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/core/api/v2/bills/{billId}": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть информацию о счете",
        "operationId": "getBillV2",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BillDtoV2"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Закрыть счет",
        "operationId": "closeBillV2",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/core/api/v1/internal/bills/{billId}": {
      "get": {
        "tags": [
          "internal"
        ],
        "summary": "Посмотреть информацию о счете пользователя",
        "description": "Этот метод предназначен для внутреннего сервиса \u0027Кредиты\u0027",
        "operationId": "getBill",
        "parameters": [
          {
            "name": "billId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "Internal Authentication": []
          }
        ]
      }
    },
...
    "/bills": {
      "get": {
        "tags": [
          "bill-controller"
        ],
        "summary": "Посмотреть все счета",
        "description": "Описание прав:\n* BILL_READ_OTHERS - доступ к любым счетам\n",
        "operationId": "getBillsV1",
        "responses": {
          "404": {
            "description": "Бизнес ошибка: что-то не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "400": {
            "description": "Бизнес ошибка или ошибка валидации",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "502": {
            "description": "Ошибка взаимодействия с другими с сервисами",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервиса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "security": [
          {
            "User Authentication": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Код ошибки. Рекомендуется использовать тип ошибки вместо кода.",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "description": "Тип ошибки",
            "enum": [
              "UNKNOWN",
              "BAD_GATEWAY",
              "CANNOT_NEGATIVE_BILL_BALANCE",
              "CLOSING_BILL_WITH_POSITIVE_BALANCE",
              "CLOSING_BILL_WITH_NEGATIVE_BALANCE",
              "TRANSACTION_WITH_ZERO_BALANCE_CHANGE",
              "CLOSING_SYSTEM_BILL",
              "BILL_NOT_FOUND",
              "CLOSED_BILL_OPERATION"
            ]
          },
          "message": {
            "type": "string",
            "description": "Сообщение об ошибке",
            "example": "Сообщение об ошибке"
          }
        }
      },
...
      "ClientBillDtoV2": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Идентификатор счета",
            "format": "uuid"
          },
          "balance": {
            "type": "integer",
            "description": "Баланс счета в копейках",
            "format": "int64"
          },
          "currency": {
            "type": "string",
            "description": "Тип валюты счета",
            "enum": [
              "RUB",
              "USD",
              "EUR",
              "JPY",
              "CNY"
            ]
          }
        }
      },
...
    }
  }
}